BackEnd/DispatcherServer/server/requirements.txt
alembic==1.13.2
antiorm==1.2.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
db==0.1.1
db-sqlite3==0.0.1
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
gunicorn
redis
requests